pyproject.toml
[project]
name = "test-agents"
version = "0.1.0"
description = "Tomi testaa agentteja ja muutenkin LLM-juttuja"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "langchain-ollama>=0.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]


